Move DataTemplateLegacy to legacy module, de-globalize _templates
